+++ CHANGELOG.md
@@ -7,6 +7,17 @@
 * `StoreFieldField` supports dot-separated field names in a bound `GridModel`, meaning it will now
   match on columns with fields such as `address.city`.
 
+### ⚙️ Technical
+
+* Note that the included version of Onsen has been replaced with a fork that includes updates for
+  react 16.13. Apps should not need to make any changes.
+
+### 📚 Libraries
+
+* react `~16.8 -> ~16.13`
+* onsenui `~16.8` -> @xh/onsenui `~16.13`
+* react-onsenui `~16.8` -> @xh/react-onsenui `~16.13`
+
 [Commit Log](https://github.com/xh/hoist-react/compare/v33.0.0...develop)
 
 ## v33.0.0 - 2020-04-22